Greyscale popup menu
I find this is less distracting, and matches what completion menus
usually look like outside of Vim. I might have to make the whole thing a
little brighter in the future because the background colour of the menu
is the same as the cursorline, which makes the two look 'connected'.